requirements.txt
fastapi
uvicorn
python-multipart
anthropic
chromadb
PyMuPDF
python-dotenv
sqlalchemy
pydantic
